Feature extractor reorganization && tests reduction part 1 && data validation fixes
