+++ interpreTS/core/features/feature_flat_spots.py
@@ -30,14 +30,6 @@ def calculate_flat_spots(data, window_size=5):
     >>> calculate_flat_spots(data)
     4
     """
-
-    # Validate the time series without requiring a DateTime index
-    validate_time_series_data(data, require_datetime_index=False, allow_nan=False)
-
-    # Check if data is empty
-    if len(data) == 0:
-        raise ValueError("Input data is empty.")
-
     run_lengths = [
         len(list(group))
         for i in range(0, len(data), window_size)
