+++ interpreTS/core/features/feature_mean.py
@@ -30,8 +30,5 @@ def calculate_mean(data):
     >>> calculate_mean(data)
     3.0
     """
-    # Validate the time series without requiring a DateTime index
-    validate_time_series_data(data, require_datetime_index=False, allow_nan=False)
-    
     # Calculate and return the mean, handling empty series by returning NaN
     return data.mean() if len(data) > 0 else np.nan
