+++ interpreTS/core/features/feature_peak.py
@@ -39,10 +39,6 @@ def calculate_peak(data, start=None, end=None):
     >>> calculate_peak(data, start=1, end=3)
     5.0
     """
-
-    # Validate the time series without requiring a DateTime index
-    validate_time_series_data(data, require_datetime_index=False)
-
     # Slice the data based on start and end, if provided
     if end is None:
         end = len(data)
