+++ interpreTS/core/features/feature_significant_changes.py
@@ -16,9 +16,6 @@ def calculate_significant_changes(data):
     float
         The proportion of significant changes in the window, in the range [0, 1].
     """
-    # Validate the time series data with allow_nan=False
-    validate_time_series_data(data, require_datetime_index=False, allow_nan=False)
-    
     # Convert np.ndarray to pd.Series
     if isinstance(data, np.ndarray):
         data = pd.Series(data)
