+++ interpreTS/core/features/mean_change.py
@@ -43,17 +43,6 @@ def calculate_mean_change(data, window_size=5):
     9     1.00
     dtype: float64
     """
-    # Validate the time series data
-    validate_time_series_data(data, require_datetime_index=False, allow_nan=False)
-
-    # Validate window_size
-    if not isinstance(window_size, int) or window_size <= 0:
-        raise ValueError("Window size must be a positive integer.")
-
-    # Handle empty or insufficient data
-    if len(data) < window_size + 1:
-        raise ValueError("The time series is too short for the specified rolling window size.")
-
     # Convert data to a pandas Series if it's an ndarray
     if isinstance(data, np.ndarray):
         data = pd.Series(data)
