+++ interpreTS/core/features/seasonality_strength.py
@@ -38,9 +38,6 @@ def calculate_seasonality_strength(data, period=2, max_lag=12):
     >>> calculate_seasonality_strength(data, period=3)
     0.75
     """
-    # Validate the time series data
-    validate_time_series_data(data, require_datetime_index=False)
-
     # Check for a valid period
     if period <= 0:
         raise ValueError("Period must be a positive integer.")
